+++ conans/test/integration/build_requires/test_toolchain_packages.py
@@ -2,7 +2,6 @@ import json
 import textwrap
 
 from conans.model.package_ref import PkgReference
-from conans.test.assets.genconanfile import GenConanfile
 from conans.test.utils.tools import TestClient
 
 
@@ -389,31 +388,3 @@ def test_compiler_gcc():
     }
     revision_dict = pkgs_json["Local Cache"]["gcc/0.1"]["revisions"][pref.ref.revision]
     assert settings_target == revision_dict["packages"][pref.package_id]["info"]["settings_target"]
-
-
-def test_per_package_settings_target():
-    c = TestClient()
-    gcc = textwrap.dedent("""
-        from conan import ConanFile
-        from conan.tools.files import save, copy
-        class Pkg(ConanFile):
-            name = "gcc"
-            version = "0.1"
-            settings = "arch"
-
-            def configure(self):
-                self.settings_target.rm_safe("os")
-                self.settings_target.rm_safe("compiler")
-                self.settings_target.rm_safe("build_type")
-
-            def package_id(self):
-                self.info.settings_target = self.settings_target
-        """)
-    c.save({"gcc/conanfile.py": gcc,
-            "app/conanfile.py": GenConanfile("app", "0.1").with_tool_requires("gcc/0.1")})
-    c.run("create gcc -s:b arch=x86_64 -s:h arch=armv7 --build-require")
-    pkg_id = c.created_package_id("gcc/0.1")
-
-    c.run("install app -s arch=armv8 -s:h gcc*:arch=armv7")
-    # it will not fail due to armv8, but use the binary for armv7
-    c.assert_listed_binary({"gcc/0.1": (pkg_id, "Cache")}, build=True)
