+++ conans/test/integration/settings/per_package_settings_test.py
@@ -95,3 +95,51 @@ class PerPackageSettingTest(unittest.TestCase):
         assert "pkg1/0.1: I am a Windows pkg!!!" in client.out
         assert "pkg2/0.1@user: I am a Linux pkg!!!" in client.out
         assert "pkg3/0.1@user/channel: I am a Linux pkg!!!" in client.out
+
+
+def test_per_package_settings_target():
+    c = TestClient()
+    gcc = textwrap.dedent("""
+        from conan import ConanFile
+
+        class Pkg(ConanFile):
+            name = "gcc"
+            version = "0.1"
+            settings = "arch"
+
+            def configure(self):
+                self.settings_target.rm_safe("os")
+                self.settings_target.rm_safe("compiler")
+                self.settings_target.rm_safe("build_type")
+
+            def package_id(self):
+                self.info.settings_target = self.settings_target
+        """)
+    c.save({"conanfile.py": gcc})
+    c.run("create . -s:b arch=x86_64 -s:h arch=armv7 --build-require")
+    pkg_id = c.created_package_id("gcc/0.1")
+
+    c.run("install --tool-requires=gcc/0.1 -s arch=armv8 -s:h gcc*:arch=armv7")
+    # it will not fail due to armv8, but use the binary for armv7
+    c.assert_listed_binary({"gcc/0.1": (pkg_id, "Cache")}, build=True)
+
+
+def test_per_package_settings_build():
+    c = TestClient()
+    cmake = textwrap.dedent("""
+        from conan import ConanFile
+        
+        class Pkg(ConanFile):
+            name = "cmake"
+            version = "0.1"
+            settings = "build_type"
+
+            def build(self):
+                self.output.info(f"Building myself with {self.settings.build_type}!!")
+        """)
+    c.save({"conanfile.py": cmake})
+    c.run("export .")
+
+    c.run("install --tool-requires=cmake/0.1 -s:b build_type=Release -s:b cmake*:build_type=Debug "
+          "--build=missing")
+    assert "cmake/0.1: Building myself with Debug!!" in c.out
